requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
python-docx==1.1.0
PyPDF2==3.0.1
openpyxl==3.1.2
Pillow==10.1.0
pandas==2.1.4
